+++ src/steps.rs
@@ -2176,6 +2176,7 @@ impl StepAccumulator {
                             uri: uri.clone(),
                             mime_type: mime_type.clone(),
                             resolution: resolution.clone(),
+                            processing: None,
                         });
                     }
                 }